Consolidate site data in one place.
