+++ lib/sources.js
@@ -1,8 +1,7 @@
 const fs = require('fs');
 const path = require('path');
-const { compose, evolve, filter, map, mergeRight } = require('ramda');
+const { assoc, compose, evolve, filter, map, mergeRight } = require('ramda');
 const nav = require('./navigation');
-const sourceRepos = require('../source-repos');
 
 const validateSourceRepo = (source = {}) => {
   const { name, mkdocs, baseURI, remote } = source;
@@ -28,31 +27,28 @@ const normalizeSourceRepo = config => evolve({
   children: compose(map(normalizeSourceRepo), filter(validateSourceRepo)),
 }, withDefaults(config));
 
-const sources = {
-  name: 'content',
-  title: 'farmOS.org',
-  baseURI: '/',
-  mkdocs: 'src/content/config.yml',
-  children: map(normalizeSourceRepo, sourceRepos),
-};
-
-const findRepoConfig = (name, repos = []) => {
-  const [head, ...tail] = repos;
+const findSourceConfig = (name, sources = []) => {
+  const [head, ...tail] = sources;
   if (!head || head.name === name) {
     return head;
   }
   if (Array.isArray(head.children)) {
-    const config = findRepoConfig(name, head.children);
+    const config = findSourceConfig(name, head.children);
     if (config) return config;
   }
   if (tail.length > 0) {
-    const config = findRepoConfig(name, tail);
+    const config = findSourceConfig(name, tail);
     if (config) return config;
   }
   return undefined;
 };
 
-function cacheSourceData() {
+const cacheSourceData = (localSources, remoteSources) => {
+  const sources = assoc(
+    'children',
+    map(normalizeSourceRepo, remoteSources),
+    localSources,
+  );
   const navigation = nav.fromSources(sources);
   const json = JSON.stringify({ sources, navigation });
   const jsonPath = '.cache/__farmOS__source_data.json';
@@ -60,6 +56,6 @@ function cacheSourceData() {
 };
 
 module.exports = {
-  findRepoConfig,
+  findSourceConfig,
   cacheSourceData,
 };
